ref: Remove EMS\CommonBundle\Common\Standard\Json
