+++ EMS/helpers/tests/Unit/Standard/JsonTest.php
@@ -93,4 +93,21 @@ class JsonTest extends TestCase
         Json::normalize($provided);
         self::assertSame($expected, Json::encode($provided));
     }
+
+    public function testIsJson(): void
+    {
+        $this->assertTrue(Json::isJson('null'));
+        $this->assertTrue(Json::isJson('54'));
+        $this->assertTrue(Json::isJson('"Foobar"'));
+        $this->assertTrue(Json::isJson('{"foo":"bar"}'));
+        $this->assertTrue(Json::isJson('[{"foo":"bar"},{"foo":"bar"}]'));
+    }
+
+    public function testIsNotJson(): void
+    {
+        $this->assertFalse(Json::isJson('FOOBAR'));
+        $this->assertFalse(Json::isJson('Foobar'));
+        $this->assertFalse(Json::isJson('{"foo":"bar"'));
+        $this->assertFalse(Json::isJson('[{"foo":"bar"}{"toto":"tata"}]'));
+    }
 }
